+++ src/main/kotlin/com/collegeapp/models/requests/InsertBookRequest.kt
@@ -3,5 +3,5 @@ package com.collegeapp.models.requests
 data class InsertBookRequest(
     val bookName: String,
     val libraryBookNumber: Long,
-    val daysAllowed: Int
+    val daysAllowed: Long
 )
